fix: circular inputs
