+++ api/admin/__init__.py
@@ -1,3 +1,4 @@
+from flask import redirect, render_template
 from flask_login import current_user
 
 from api.utils import Role
